Fix: moving the data to models and changing some imports
