+++ src/app/page.tsx
@@ -2,12 +2,12 @@ import Header from "@/components/organism/header";
 import Hero from "@/components/organism/sections/home/hero";
 import CompletionTools from "@/components/organism/sections/home/completionTools";
 import ImageChanger from "@/components/organism/sections/home/iteraction";
-import { data } from "@/data/iteraction-data/data";
-import { data2 } from "@/data/sales-data/data";
 import { Automated } from "@/components/organism/sections/home/automated";
 import CompletionProcess from "@/components/organism/sections/home/completionProcess";
 import { InstantChanges } from "@/components/organism/sections/home/instantChanges";
 import { WhySalesAssistSalesProcess } from "@/components/organism/sections/home/whySalesAssistSales";
+import { actionsData } from "@/model/api/image-data/data";
+import { industriesData } from "@/model/api/sales-data/data";
 
 
 export default function Home() {
@@ -16,11 +16,11 @@ export default function Home() {
       <Header/>
       <Hero/>
       <CompletionTools/>
-      <ImageChanger data={data}/>
+      <ImageChanger data={actionsData}/>
       <Automated></Automated>
-      <CompletionProcess></CompletionProcess>
-      <InstantChanges></InstantChanges>
-      <WhySalesAssistSalesProcess data={data2} ></WhySalesAssistSalesProcess>
+      <CompletionProcess/>
+      <InstantChanges/>
+      <WhySalesAssistSalesProcess data={industriesData} />
     </main>
   )
 }
