+++ src/components/atoms/salesIcon/index.tsx
@@ -1,4 +1,4 @@
-import Image, { ImageProps } from "next/image";
+import Image from "next/image";
 
 interface Icon {
     code: string,
